birthday_reminder_backend/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.3.1
